Schema Fields:
The schema defines the fields (properties) for the User model:
fullname:
Type: String
Required: true
This field stores the full name of the user.
email:
Type: String
Required: true
This field stores the email address of the user.
password:
Type: String
Required: true
This field stores the hashed password of the user.
profilePhoto:
Type: String
Default: ""
This field stores the URL of the user's profile photo. It defaults to an empty string if not provided.
Timestamps Option:

The schema is created with an additional option { timestamps: true }.
This option automatically adds two fields to the schema:
createdAt:
This field stores the date and time when the document was created.
updatedAt:
This field stores the date and time when the document was last updated.