Importing Dependencies Express:
The express module is imported at the beginning of the file. Express is a web application framework for Node.js, which provides robust features for building web and mobile applications.
User Controller Functions:

The login, logout, and register functions are imported from the userController.js file. These functions handle the logic for user authentication and registration.