Defining Routes Register Route:
router.route("/register").post(register);
This line sets up a POST route at the /register endpoint.
When a POST request is made to /register, the register function from the controller is invoked to handle the request. This function typically handles user registration logic such as input validation, password hashing, and storing the new user in the database.
Login Route:

router.route("/login").post(login);
This line sets up a POST route at the /login endpoint.
When a POST request is made to /login, the login function from the controller is invoked to handle the request. This function usually handles user authentication logic such as verifying the email and password, and generating a JWT for the user.
Logout Route:

router.route("/logout").get(logout);
This line sets up a GET route at the /logout endpoint.
When a GET request is made to /logout, the logout function from the controller is invoked to handle the request. This function typically handles the logic for logging out a user, such as clearing the authentication cookie.