Schema Fields:
The schema defines the fields (properties) for the Email model:
to:
Type: String
Required: true
This field stores the recipient's email address.
subject:
Type: String
Required: true
This field stores the subject of the email.
message:
Type: String
Required: true
This field stores the body of the email.
userId:
Type: mongoose.Schema.Types.ObjectId
Reference (ref): 'User'
This field stores the ID of the user who sent the email. It is an ObjectId that references the User model, creating a relationship between the email and the user.