Importing Dependencies Express:
The express module is imported at the beginning of the file. Express is a web application framework for Node.js that simplifies the process of building web applications and APIs.
Email Controller Functions:

The createEmail, deleteEmail, and getAllEmailById functions are imported from the emailController.js file. These functions handle the business logic for creating, deleting, and retrieving emails.
Authentication Middleware:

The isAuthenticated middleware is imported from the isAuthenticated.js file. This middleware ensures that routes can only be accessed by authenticated users.