Conditional Rendering:
If authUser is not present:
Renders the RouterProvider with the defined routes, which includes the Login and Signup pages.
If authUser is present:
Renders the Navbar component for navigation.
Renders the RouterProvider with the defined routes for authenticated users.
Renders the SendEmail component at the bottom right of the screen for composing new emails.